LayoutTests/streams/readable-stream.html

<!DOCTYPE html>
<script src='../resources/testharness.js'></script>
<script src='../resources/testharnessreport.js'></script>
<script src='../resources/gc.js'></script>
<script>
test(function() {
    assert_throws(new TypeError(), function() {
        new ReadableStream(null);
    }, 'constructor should throw when the source is null');
}, 'ReadableStream can\'t be constructed with garbage');

test(function()
{
    var isStartCalled = false;
    var source = {
        start: function(controller) {
            assert_equals(this, source, 'source is this during start');

            // FIXME: We should add constructor at some point.
            var methods = [ 'close', 'enqueue', 'error' ];
            var proto = Object.getPrototypeOf(controller);

            assert_array_equals(Object.getOwnPropertyNames(Object.getPrototypeOf(controller)).sort(), methods,
                        'the controller should have the right methods');

            for (var m of methods) {
                var methodProperties = [ 'length', 'name' ];
                var propDesc = Object.getOwnPropertyDescriptor(proto, m);
                assert_equals(propDesc.enumerable, true, 'method should be enumerable');
                assert_equals(propDesc.configurable, true, 'method should be configurable');
                assert_equals(propDesc.writable, true, 'method should be writable');
                assert_equals(typeof controller[m], 'function', 'should have be a method');
                assert_array_equals(Object.getOwnPropertyNames(controller[m]).sort(), methodProperties, 'method should have the right properties');
            }

            assert_equals(controller.close.length, 0, 'close should have no parameters');
            assert_equals(controller.enqueue.length, 1, 'enqueue should have 1 parameter');
            assert_equals(controller.error.length, 1, 'error should have 1 parameter');
            //assert_equals(controller.constructor.length, 1, 'constructor should have 1 parameters');

            isStartCalled = true;
        }
    };

    var rs = new ReadableStream(source);
    assert_true(isStartCalled);
}, 'ReadableStream start should be called with the proper parameters');

test(function()
{
    var isStartCalled = false;

    var SimpleStreamSource = function() { };
    SimpleStreamSource.prototype.start = function() { isStartCalled = true; };
    SimpleStreamSource.prototype.constructor = SimpleStreamSource;

    var rs = new ReadableStream(new SimpleStreamSource());
    assert_true(isStartCalled);
}, 'ReadableStream should be able to call start method within prototype chain of its source');

t1 = async_test('A readable stream controller methods should throw if its readablestream is collected');
t1.step(function() {
    var controller;
    new ReadableStream({
        start: function(c) {
            controller = c;
        }
    });
    setTimeout(t1.step_func(function() {
        window.gc();
        assert_throws(new TypeError(), function() { controller.close(); });
        assert_throws(new TypeError(), function() { controller.error(); });
        assert_throws(new TypeError(), function() { controller.enqueue(); });
        t1.done();
    }), 10);
});
</script>
